Finished NavBar and created a spacer component to separate buttons easily
