+++ frontend/src/components/ButtonSpacer/ButtonSpacer.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+
+const ButtonSpacer = () => 
+    {
+        return(
+            <div>
+            &nbsp;&nbsp;&nbsp;
+        </div>
+        )
+    };
+
+
+export default ButtonSpacer
