+++ frontend/src/components/LandingPage/LandingPage.js
@@ -1,12 +1,11 @@
 import React from 'react'
-import MyNavbar from './MyNavbar/MyNavbar';
+import MyNavbar from '../MyNavbar/MyNavbar';
 //Import components that you create here
 
 function LandingPage(){
     return(
         //Use components that you import here
         <>
-            
              <MyNavbar/>
         </>
